Split workers.ts
