+++ api/src/index.dev.ts
@@ -1,6 +1,6 @@
 import wasmModule from 'kanatrans-wasm/debug/kanatrans-wasm.wasm';
 
-import { workers } from './workers';
+import { worker } from '@/worker';
 
 // biome-ignore lint/style/noDefaultExport:
-export default workers(wasmModule);
+export default worker(wasmModule);
