+++ api/src/index.ts
@@ -1,5 +1,6 @@
 import wasmModule from 'kanatrans-wasm/kanatrans-wasm.wasm';
-import { workers } from './workers';
+
+import { worker } from '@/worker';
 
 // biome-ignore lint/style/noDefaultExport:
-export default workers(wasmModule);
+export default worker(wasmModule);
