+++ api/src/types/output.ts
@@ -1,0 +1,3 @@
+export type OutputStream = TransformStream<Uint8Array<ArrayBufferLike>, Uint8Array<ArrayBufferLike>>;
+
+export type OutputStreamReadResult = ReadableStreamReadResult<Uint8Array<ArrayBufferLike>>;
