+++ api/src/types/workers.ts
@@ -1,0 +1,2 @@
+// biome-ignore lint/suspicious/noEmptyInterface:
+export interface Env {}
