+++ api/tsconfig.json
@@ -2,6 +2,9 @@
   "extends": ["../tsconfig.json"],
   "compilerOptions": {
     "types": ["@cloudflare/workers-types"],
-    "outDir": "dist"
+    "outDir": "dist",
+    "paths": {
+      "@/*": ["./src/*"]
+    }
   }
 }
